+++ examples/simple.rs
@@ -1,4 +1,7 @@
-use progressing::{self, Bar};
+use progressing::{
+    bernoulli::Bar as BernoulliBar, clamping::Bar as ClampingBar, mapping::Bar as MappingBar,
+    Baring,
+};
 
 fn main() {
     // different examples for different use-cases
@@ -17,7 +20,7 @@ fn main() {
 /// [=====>------------]
 fn clamped() {
     println!("Printing value 0.3 clamped to [0, 1]");
-    let mut progress_bar = progressing::ClampingBar::new();
+    let mut progress_bar = ClampingBar::new();
     progress_bar.set_len(20);
     progress_bar.set(0.3);
     println!("{}", progress_bar);
@@ -27,7 +30,7 @@ fn clamped() {
 /// [================>-] (4 / 5)
 fn mapped() {
     println!("Mapping from [-9, 5] to [0, 1]");
-    let mut progress_bar = progressing::MappingBar::new(-9, 5);
+    let mut progress_bar = MappingBar::with_range(-9, 5);
     progress_bar.set_len(20);
     progress_bar.set(4);
     println!("{}", progress_bar);
@@ -37,7 +40,7 @@ fn mapped() {
 /// [============>-----] (42 / 60 # 130)
 fn bernoulli() {
     println!("Bernoulli-Bar counting successes (42 / 60) and attempts (# 130)");
-    let mut progress_bar = progressing::BernoulliBar::from_goal(60);
+    let mut progress_bar = BernoulliBar::with_goal(60);
     progress_bar.set_len(20);
     progress_bar.set((42, 130));
     println!("{}", progress_bar);
@@ -47,8 +50,7 @@ fn bernoulli() {
 /// [================>-] (4 / 5) ~ 2 min
 fn timed_mapped() {
     println!("Mapping from [-9, 5] to [0, 1], but with time-approximation");
-    let progress_bar = progressing::MappingBar::new(-9, 5);
-    let mut progress_bar = progressing::TimedBar::new(progress_bar);
+    let mut progress_bar = MappingBar::with_range(-9, 5).timed();
     progress_bar.set_len(20);
     progress_bar.set(4);
     println!("{}", progress_bar);
@@ -57,7 +59,7 @@ fn timed_mapped() {
 // clamped-example, but with other styles
 fn styles() {
     println!("Custom styles");
-    let mut progress_bar = progressing::ClampingBar::new();
+    let mut progress_bar = ClampingBar::new();
     progress_bar.set_len(20);
     progress_bar.set(0.3);