+++ src/bernoulli.rs
@@ -1,43 +1,49 @@
-use crate::{bars::mapping, Bar, MappingBar};
+use crate::{mapping, timing, Baring};
 use std::{
     fmt::{self, Display},
     ops::{Add, AddAssign, Div, Sub},
 };
+
+pub struct Config {}
 
 /// A progress-bar counting successes (e.g. `42 out of 60`) and respective attempts (e.g. `130`).
 ///
 /// # Mini-Example
 ///
 /// ```
-/// use progressing::Bar;
+/// use progressing::{Baring, bernoulli::Bar as BernoulliBar};
 ///
 /// /// Bernoulli-Bar counting successes (42 / 60) and attempts (# 130)
 /// /// [============>-----] (42 / 60 # 130)
 /// fn main() {
 ///     println!("Bernoulli-Bar counting successes (42 / 60) and attempts (# 130)");
-///     let mut progress_bar = progressing::BernoulliBar::from_goal(60);
+///     let mut progress_bar = BernoulliBar::with_goal(60);
 ///     progress_bar.set_len(20);
 ///     progress_bar.set((42, 130));
 ///     println!("{}", progress_bar);
 /// }
 /// ```
 #[derive(Debug)]
-pub struct BernoulliBar {
-    bar: MappingBar<usize>,
+pub struct Bar {
+    bar: mapping::Bar<usize>,
     attempts: usize,
 }
 
-impl BernoulliBar {
-    pub fn from_goal(n: usize) -> BernoulliBar {
-        BernoulliBar {
-            bar: MappingBar::new(0, n),
+impl Bar {
+    pub fn with_goal(end: usize) -> Bar {
+        Bar {
+            bar: mapping::Bar::with_range(0, end),
             attempts: 0,
         }
     }
+
+    pub fn timed(self) -> timing::Bar<Bar> {
+        timing::Bar::with(self)
+    }
 }
 
-impl Bar for BernoulliBar {
-    type Progress = BernoulliProgress;
+impl Baring for Bar {
+    type Progress = Progress;
 
     fn len(&self) -> usize {
         self.bar.len()
@@ -47,8 +53,8 @@ impl Bar for BernoulliBar {
         self.bar.set_len(new_bar_len)
     }
 
-    fn progress(&self) -> BernoulliProgress {
-        BernoulliProgress {
+    fn progress(&self) -> Progress {
+        Progress {
             successes: self.bar.progress(),
             attempts: self.attempts,
         }
@@ -56,43 +62,43 @@ impl Bar for BernoulliBar {
 
     fn set<P>(&mut self, outcome: P)
     where
-        P: Into<BernoulliProgress>,
+        P: Into<Progress>,
     {
         let outcome = outcome.into();
         self.bar.set(outcome.successes);
         self.attempts = outcome.attempts;
     }
 
-    fn start(&self) -> BernoulliProgress {
-        BernoulliProgress {
+    fn start(&self) -> Progress {
+        Progress {
             successes: self.bar.start(),
             attempts: 0,
         }
     }
 
-    fn end(&self) -> BernoulliProgress {
-        BernoulliProgress {
+    fn end(&self) -> Progress {
+        Progress {
             successes: self.bar.end(),
             attempts: 1,
         }
     }
 
-    fn has_progressed_much(&self) -> bool {
-        self.bar.has_progressed_much()
+    fn has_progressed_significantly(&self) -> bool {
+        self.bar.has_progressed_significantly()
     }
 
-    fn remember_progress(&mut self) {
-        self.bar.remember_progress()
+    fn remember_significant_progress(&mut self) {
+        self.bar.remember_significant_progress()
     }
 }
 
-impl Display for BernoulliBar {
+impl Display for Bar {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
             f,
             "{} ({} / {} # {})",
-            mapping::inner_bar(&self.bar),
-            mapping::inner_k(&self.bar),
+            self.bar,
+            self.bar.progress(),
             self.bar.end(),
             self.attempts
         )
@@ -101,70 +107,70 @@ impl Display for BernoulliBar {
 
 /// Just a simple struct capsuling access to successes and attempts.
 #[derive(Copy, Clone)]
-pub struct BernoulliProgress {
+pub struct Progress {
     pub successes: usize,
     pub attempts: usize,
 }
 
-impl From<(usize, usize)> for BernoulliProgress {
-    fn from((successes, attempts): (usize, usize)) -> BernoulliProgress {
-        BernoulliProgress {
+impl From<(usize, usize)> for Progress {
+    fn from((successes, attempts): (usize, usize)) -> Progress {
+        Progress {
             successes,
             attempts,
         }
     }
 }
 
-impl From<usize> for BernoulliProgress {
-    fn from(successes: usize) -> BernoulliProgress {
-        BernoulliProgress {
+impl From<usize> for Progress {
+    fn from(successes: usize) -> Progress {
+        Progress {
             successes,
             attempts: successes,
         }
     }
 }
 
-impl From<bool> for BernoulliProgress {
-    fn from(is_successful: bool) -> BernoulliProgress {
-        BernoulliProgress {
+impl From<bool> for Progress {
+    fn from(is_successful: bool) -> Progress {
+        Progress {
             successes: if is_successful { 1 } else { 0 },
             attempts: 1,
         }
     }
 }
 
-impl Add for BernoulliProgress {
-    type Output = BernoulliProgress;
+impl Add for Progress {
+    type Output = Progress;
 
-    fn add(self, other: BernoulliProgress) -> BernoulliProgress {
-        BernoulliProgress {
+    fn add(self, other: Progress) -> Progress {
+        Progress {
             successes: self.successes + other.successes,
             attempts: self.attempts + other.attempts,
         }
     }
 }
 
-impl AddAssign for BernoulliProgress {
-    fn add_assign(&mut self, other: BernoulliProgress) {
+impl AddAssign for Progress {
+    fn add_assign(&mut self, other: Progress) {
         *self = *self + other;
     }
 }
 
-impl Sub for BernoulliProgress {
-    type Output = BernoulliProgress;
+impl Sub for Progress {
+    type Output = Progress;
 
-    fn sub(self, subtrahend: BernoulliProgress) -> BernoulliProgress {
-        BernoulliProgress {
+    fn sub(self, subtrahend: Progress) -> Progress {
+        Progress {
             successes: self.successes - subtrahend.successes,
             attempts: self.attempts - subtrahend.attempts,
         }
     }
 }
 
-impl Div for BernoulliProgress {
+impl Div for Progress {
     type Output = f64;
 
-    fn div(self, dividend: BernoulliProgress) -> f64 {
+    fn div(self, dividend: Progress) -> f64 {
         self.successes as f64 / (dividend.successes as f64)
     }
 }